+++ demo/app/components/editor-layer/polygons-editor-layer/polygons-editor-layer.component.ts
@@ -1,9 +1,9 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { PolygonEditUpdate } from '../../../../../src/angular-cesium-widgets/models/polygon-edit-update';
-import { EditActions } from '../../../../../src/angular-cesium-widgets/models/edit-actions.enum';
-import { PolygonEditorObservable } from '../../../../../src/angular-cesium-widgets/models/polygon-editor-observable';
-import { PolygonsEditorService } from '../../../../../src/angular-cesium-widgets/services/entity-editors/polygons-editor/polygons-editor.service';
-import { LabelProps } from '../../../../../src/angular-cesium-widgets/models/label-props';
+import {ChangeDetectionStrategy, Component, OnInit} from '@angular/core';
+import {PolygonEditUpdate} from '../../../../../src/angular-cesium-widgets/models/polygon-edit-update';
+import {EditActions} from '../../../../../src/angular-cesium-widgets/models/edit-actions.enum';
+import {PolygonEditorObservable} from '../../../../../src/angular-cesium-widgets/models/polygon-editor-observable';
+import {PolygonsEditorService} from '../../../../../src/angular-cesium-widgets/services/entity-editors/polygons-editor/polygons-editor.service';
+import {LabelProps} from '../../../../../src/angular-cesium-widgets/models/label-props';
 
 @Component({
   selector: 'polygons-editor-layer',
@@ -26,6 +26,7 @@ export class PolygonsEditorLayerComponent implements OnInit {
     if (this.editing$) {
       this.stopEdit();
     }
+    // create accepts PolygonEditOptions object
     this.editing$ = this.polygonsEditor.create();
     this.editing$.subscribe((editUpdate: PolygonEditUpdate) => {
       console.log(editUpdate.points); // point = position with id